weather_rest_api/requirements.txt
flask
flask_restful
mysql-connector-python
requests
dotenv